Add files via upload
This program shows a LIVE Stock price window for Stock listed on NSE. This window updates in every 5 seconds and it appears at the bottom-right of Computer screen, So you can look at live price easily. An alert sound rings if Price crosses upword or downword accoring to '+' or '-' sign in the end. For e.g. If you enter "RELIANCE 2000+", alert sound will ring if it crosses 2k on NSE index. vice versa for negative values.  
This is purely for education purpose not commercial.